refactor: split compile_time_continuation.go by domain (4.1)
Split 2,374-line file into four domain-specific files:
- compile_time_continuation.go (433 lines): core struct, symbol
  resolution, procedure calls, expression compilation
- compile_time_continuation_quasiquote.go (565 lines): quasiquote
  expansion, unquote handling, needs-runtime checks
- compile_time_continuation_library.go (1,201 lines): define-library,
  import/export, cond-expand, define-syntax, import set parsing
- compile_time_continuation_include.go (247 lines): include/include-ci,
  file resolution, letrec* semantics

Pure organizational split — no logic changes.
